requirements.txt
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.2.1
Automat==0.8.0
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
cffi==1.13.2
channels==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
daphne==2.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2
django-allauth==0.41.0
django-cors-headers==3.2.1
django-cors-middleware==1.5.0
django-countries==6.1.2
django-crispy-forms==1.7.0
django-debug-toolbar==2.2
django-extensions==2.2.8
django-material==0.5.1
django-rest-auth==0.9.5
django-restframework==0.0.1
django-tinymce4-lite==1.7.5
django-webpack-loader==0.7.0
djangorestframework==3.11.0
docopt==0.6.2
get==2019.4.13
gitignore==0.0.8
gunicorn==20.0.4
html5lib==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
jsmin==2.2.2
linkedin-scraper==2.4.3
lxml==4.5.0
mysqlclient==1.4.4
numpy==1.18.2
oauthlib==3.1.0
package-name==0.1
pandas==1.0.3
parsel==1.5.2
Pillow==6.2.0
post==2019.4.13
Protego==0.1.16
proxyscrape==0.3.0
psycopg2==2.8.4
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==2.0.0
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.2
#pywin32==227
query-string==2019.4.13
queuelib==1.5.0
request==2019.4.13
requests==2.22.0
requests-oauthlib==1.3.0
Scrapy==1.8.0
scrapy-proxy-pool==0.1.9
scrapy-user-agents==0.1.1
selenium==3.141.0
service-identity==18.1.0
six==1.14.0
soupsieve==1.9.5
South==1.0.2
sqlparse==0.3.0
static3==0.7.0
stripe==2.44.0
Twisted==19.10.0
txaio==20.1.1
ua-parser==0.10.0
urllib3==1.25.8
user-agents==2.1
virtualenv==16.7.4
virtualenvwrapper-win==1.2.5
w3lib==1.21.0
webencodings==0.5.1
whitenoise==5.0.1
zope.interface==4.7.1
